# src/swaggerOutput.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "TradeMark",
    "description": "TradeMark REST API Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/",
      "description": "Local Host"
    }
  ],
  "paths": {
...
    "/admin/earnings/all": {
      "get": {
        "tags": ["Admin - Earnings"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarningssResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/EarningssResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/earnings/user/{uid}": {
      "get": {
        "tags": ["Admin - Earnings"],
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EarningssResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/EarningssResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/earnings/{uid}": {
      "post": {
        "tags": ["Admin - Earnings"],
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Earnings added to the user balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "406": {
            "description": "Not Acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidID"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EarningRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/EarningRequest"
              }
            }
          }
        }
      }
    },
    "/admin/earnings/deduct/{uid}": {
      "post": {
        "tags": ["Admin - Earnings"],
        "description": "",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Earnings deducted from the user balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/NotExists"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EarningRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/EarningRequest"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EarningRequest": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string",
            "example": "number"
          },
          "plan_id": {
            "type": "string",
            "example": "string"
          },
          "investment_id": {
            "type": "string",
            "example": "string"
          }
        },
        "required": ["amount", "plan_id", "investment_id"],
        "xml": {
          "name": "EarningRequest"
        }
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "string"
          }
        },
        "xml": {
          "name": "Response"
        }
      },
...
      "BadRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Bad Request."
          }
        },
        "xml": {
          "name": "BadRequest"
        }
      },
      "InvalidToken": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid Token."
          }
        },
        "xml": {
          "name": "InvalidToken"
        }
      },
...
      "NotExists": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Does not exists."
          }
        },
        "xml": {
          "name": "NotExists"
        }
      },
      "InvalidID": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid ID."
          }
        },
        "xml": {
          "name": "InvalidID"
        }
      },
...
    }
  },
...
}